IMPROVE: Update name of Mon
[Description]
Zoning adjuster is already taken by another Modder. I was silly
enough to think that the name isn't taken. But being one of my favourite
mods from CS 1, the name Zoning Adjuster was in my mind. Just didn't
remember that it was already taken. This change is to rename the mod
to ZoningToolkit as discussed with the other Modder. Going forward
this Mod will be called ZoningToolkit.